requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
pydantic>=1.8.0
python-dotenv>=0.19.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
